feat(prompts): add prompts as first class citizens (#145)
Adds Prompts as first class citizens. This is a breaking change as
SimplePrompt with just a a `&str` is no longer allowed.

This introduces `Prompt` and `PromptTemplate`. A template uses jinja
style templating build on tera. Templates can be converted into prompts,
and have context added. A prompt is then send to something that prompts,
i.e. openai or bedrock.

Additional prompts can be added either compiled or as one-offs.
Additionally, it's perfectly fine to prompt with just a string as well,
just provide an `.into()`.

For future development, some LLMs really benefit from system prompts,
which this would enable. For the query pipeline we can also take a much
more structured approach with composed templates and conditionals.