feat(indexing): Metadata as first class citizen (#204)
Adds our own implementation for metadata, internally still using a
BTreeMap. The Value type is now a `serde_json::Value` enum. This allows
us to store the metadata in the same format as the rest of the document,
and also allows us to use values programmatically later.

As is, all current meta data is still stored as Strings.

Closes #162