feat(indexing)!: Default LLM for indexing pipeline and boilerplate Transformer macro (#227)
Add setting a default LLM for an indexing pipeline, avoiding the need to
clone multiple times.

More importantly, introduced `swiftide-macros` with
`#[swiftide_macros::indexing_transformer]` that generates
all boilerplate code used for internal transformers. This ensures all
transformers are consistent and makes them
easy to change in the future. This is a big win for maintainability and
ease to extend. Users are encouraged to use the macro
as well.

BREAKING CHANGE: Introduces `WithIndexingDefaults` and
`WithBatchIndexingDefaults` trait constraints for transformers. They can
be used as a marker
with a noop (i.e. just `impl WithIndexingDefaults for MyTransformer
{}`). However, when implemented fully, they can be used to provide
defaults from the pipeline to your transformers.