feat: Major performance improvements (#291)
Futures that do not yield were not run in parallel properly. With this
futures are spawned on a tokio worker thread by default.

When embedding (fastembed) and storing a 85k row dataset, there's a
~1.35x performance improvement:
<img width="621" alt="image"
src="https://github.com/user-attachments/assets/ba2d4d96-8d4a-44f1-b02d-6ac2af0cedb7">

~~Need to do one more test with IO bound futures as well. Pretty huge,
not that it was slow.~~

With IO bound openai it's 1.5x.