feat(agents): **EXPERIMENTAL** Agents in Swiftide (#463)
Agents are coming to Swiftide! We are still ironing out all the kinks,
while we make it ready for a proper release. You can already experiment
with agents, see the rustdocs for documentation, and an example in
`/examples`, and feel free to contact us via github or discord. Better
documentation, examples, and tutorials are coming soon.

Run completions in a loop, define tools with two handy macros, customize
the agent by hooking in on lifecycle events, and much more.

Besides documentation, expect a big release for what we build this for
soon! 🎉