feat(query): Documents as first class citizens (#504)
For simple RAG, just adding the content of a retrieved document might be
enough. However, in more complex use cases, you might want to add
metadata as well, as is or for conditional formatting.

For instance, when dealing with large amounts of chunked code, providing
the path goes a long way. If generated metadata is good enough, could be
useful as well.

With this retrieved Documents are treated as first class citizens,
including any metadata as well. Additionally, this also paves the way
for multi retrieval (and multi modal).