docs: document feature flags using doc_auto_cfg
https://doc.rust-lang.org/rustdoc/unstable-features.html#extensions-to-the-doc-attribute

This is an unstable feature, so it's feature flagged with a feature that
is always passed by docs.rs. To run the same settings use cargo-docs-rs

Also moves the logo into the top left corner of all docs, rather than
just in the main doc.

Makes the docs.rs config consistent across all crate manifests.