refactor!: Remove templates (#716)
Template / prompt interface got confusing and bloated. This removes
`Template` fully, and changes Prompt such that it can either ref to a
one-off, or to a template named compiled in the swiftide repository.

BREAKING CHANGE: This removes `Template` from Swiftide and simplifies
the whole setup significantly. The internal Swiftide Tera repository can
still be extended like with Templates. Same behaviour with less code and
abstractions.