feat(integrations): Streaming chat completions for OpenAI (#741)
Extends the `ChatCompletion` trait with a `complete_stream` method,
which returns a stream of `ChatCompletionResponse`. If the client does
not implement streaming, a one shot is returned. Making a small
concession on performance such that each chatcompletion response also
includes the full accumulated response. Each streaming message tries to
rebuild the tools.

Definately some room for performance improvements. Might or might not do
that for an initial release.

In a future PR, we can let use this to add a query/query_once stream
variant, or always stream (lazilly) if the user has taken a stream
handle from the agent. Still need to decide on that implementation.
tl&dr; streaming is cool we can do it.

Needs tests.