feat(agents)!: Easy human-in-the-loop flows by decorating tools (#790)
Tools can now return a FeedbackRequired with an optional Value payload.
The agent will stop, with the tool call and optional payload in the
reason. Possible to use either the stop reason returned by the agent or
a hook to act on this, or use it in your own tools.

Feedback is given on the context of the agent with the tool call and an
optional payload. The payload offers flexibility to provide feedback to
the tool, or to provide context to the user.

BREAKING CHANGE: The `Tool` trait now receives a `ToolCall` as argument
instead of an `Option<&str>`. The latter is still accessible via
`tool_call.args()`.