feat(integrations/openai)!: Present and re-use reasoning in responses api (#988)
Ensures the the compacted reasoning prompt is present in follow up turns
when using agents with reasoning. Enabled by default if reasoning is
enabled.

BREAKING CHANGE: Adds a `ChatMessage::Reasoning` that will include both
the summary and compressed reasoning, so follow up turns in agents can
use it properly, and users can display the summary in the UI if they
want.